Group No Suggestions almost works.
When the algorithm returns nothing, the members are kicked back to the keywords page and the host goes to the expand radius page. Quiz will not automatically resubmit, must click 'Find Restaurant' to continue (or host can skip). When all suggestions are rejected, the same flow is used. Known issue: Old suggestions are used the second time the group gets suggestions